Rename chatterinowebsocketpplogger.hpp to ChatterinoWebSocketppLogger.hpp
This is to fit the Chatterino filename standard better
